+++ src/main/java/com/ing/data/cassandra/jdbc/AbstractResultSet.java
@@ -17,9 +17,9 @@ package com.ing.data.cassandra.jdbc;
 
 import com.datastax.oss.driver.api.core.type.DataType;
 import com.ing.data.cassandra.jdbc.types.DataTypeEnum;
-import edu.umd.cs.findbugs.annotations.NonNull;
 import org.apache.commons.lang3.StringUtils;
 
+import javax.annotation.Nonnull;
 import java.io.InputStream;
 import java.io.Reader;
 import java.math.BigDecimal;
@@ -56,7 +56,7 @@ abstract class AbstractResultSet implements Wrapper {
      * @param type        The data type to check.
      * @return {@code true} if the column CQL data type is the given one, {@code false} otherwise.
      */
-    boolean isCqlType(final int columnIndex, @NonNull final DataTypeEnum type) {
+    boolean isCqlType(final int columnIndex, @Nonnull final DataTypeEnum type) {
         final String columnType = StringUtils.substringBefore(DataTypeEnum.cqlName(getCqlDataType(columnIndex)), "<");
         return type.cqlType.equalsIgnoreCase(columnType);
     }
@@ -68,7 +68,7 @@ abstract class AbstractResultSet implements Wrapper {
      * @param type        The data type to check.
      * @return {@code true} if the column CQL data type is the given one, {@code false} otherwise.
      */
-    boolean isCqlType(final String columnLabel, @NonNull final DataTypeEnum type) {
+    boolean isCqlType(final String columnLabel, @Nonnull final DataTypeEnum type) {
         final String columnType = StringUtils.substringBefore(DataTypeEnum.cqlName(getCqlDataType(columnLabel)), "<");
         return type.cqlType.equalsIgnoreCase(columnType);
     }